consolidate triangulation widget
